This works for transmit.
Former-commit-id: 4cbd2a1974255a5e6f93f0924e6351babb0ad6b0

+++ libraries/CAN/src/CAN.ipp
@@ -358,8 +358,8 @@ bool ArduinoCAN<CAN_MTU_SIZE>::cfg_pins(int const max_index, int const can_tx_pi
   }
 
   /* Actually configure pin functions. */
-  R_IOPORT_PinCfg(&g_ioport_ctrl, g_pin_cfg[can_tx_pin].pin, (uint32_t) (IOPORT_CFG_PERIPHERAL_PIN | IOPORT_PERIPHERAL_CAN));
-  R_IOPORT_PinCfg(&g_ioport_ctrl, g_pin_cfg[can_rx_pin].pin, (uint32_t) (IOPORT_CFG_PERIPHERAL_PIN | IOPORT_PERIPHERAL_CAN));
+  R_IOPORT_PinCfg(&g_ioport_ctrl, g_pin_cfg[can_tx_pin].pin, (uint32_t)(IOPORT_CFG_PERIPHERAL_PIN) | (uint32_t)(IOPORT_PERIPHERAL_CAN));
+  R_IOPORT_PinCfg(&g_ioport_ctrl, g_pin_cfg[can_rx_pin].pin, (uint32_t)(IOPORT_CFG_PERIPHERAL_PIN) | (uint32_t)(IOPORT_PERIPHERAL_CAN));
 
   return true;
 }
